DixitGP/requirements.txt
aiogram==2.23.1
aiohttp
pydantic[dotenv]==1.10.2
pillow
requests
